+++ tests/TestEvents/WithoutCustomPayload.php
@@ -1,0 +1,14 @@
+<?php
+
+
+    declare(strict_types = 1);
+
+
+    namespace Tests\TestEvents;
+
+    class WithoutCustomPayload
+    {
+
+
+
+    }